dont delete image if default

+++ app/Http/Controllers/PostsController.php
@@ -184,7 +184,9 @@ class PostsController extends Controller
         $post->body = $request->input('body');
         if($request->hasFile('cover_image')){
             // Delete old image
-            Storage::delete('public/cover_images/'.$post->cover_image);
+            if($post->cover_image != 'noimage.jpg'){
+                Storage::delete('public/cover_images/'.$post->cover_image);
+            }
         }
         $post->cover_image = $fileNameToStore;
         $post->save();
